Turn off rc_expand_param before loading oh-my-zsh.
The upgrade to macOS Sierra caused an upgrade to ZSH. Something changed
in the Sierra ZSH (5.2) when compared to Yosemite (5.0.5):
Now, if rc_expand_param is turned on (which it is in gcv's config),
zplugin fails to load oh-my-zsh with "eval" errors. Turning it off seems
to restore the working order. TBD whether I want rc_expand_param on or
not.